+++ libs/components/src/lib/split-button/index.ts
@@ -1,0 +1,3 @@
+import { registerSplitButton } from './definition';
+
+registerSplitButton();