+++ libs/components/src/shared/localization/Locale.ts
@@ -1,0 +1,5 @@
+import type { DatePickerLocale } from '../../lib/date-picker/locale';
+
+export interface Locale {
+	datePicker: DatePickerLocale;
+}